backend/requirements.txt
requests>=2.31.0
pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
websocket-client>=1.6.0
